requirements.txt
streamlit
alpaca-py
pandas
numpy
python-dotenv
requests
